requirements.txt
setuptools
sphinx
requests
PyYAML
